Adding Defaults !requiretty to the sudoers file
Originally it started of having a sed command doing so.
It seemed that the default sudoers file does not have any match on "requiretty".
Therefore with this change it can ensure the removal of any possible matched that
may contain conflicting matches and appends a rule to disable "require tty".